it-da-ai-server/requirements.txt
# 프로젝트 루트에서
# py -3.11 -m venv venv

# 가상환경 활성화
# venv\Scripts\activate

# python -m pip install --upgrade pip
# python -m pip install -r requirements.txt

# ===============================
# Python 3.11.9 안정 조합
# ===============================

# FastAPI
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
python-multipart==0.0.9

# Core numerical (🔥 NumPy 2.x 금지)
numpy==1.26.4
pandas==2.1.4

# ML
scikit-learn==1.3.2
lightgbm==4.1.0

# NLP
torch==2.1.2
transformers==4.37.2

# HTTP
httpx==0.26.0
requests==2.31.0

# OpenAI
openai==1.10.0

# Utils
python-dotenv==1.0.1